+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_job/sweep/sampling_algorithm.py
@@ -54,7 +54,6 @@ class RandomSamplingAlgorithm(SamplingAlgorithm):
 
     .. admonition:: Example:
 
-
         .. literalinclude:: ../../../../../../samples/ml_samples_sweep_configurations.py
             :start-after: [START configure_sweep_job_random_sampling_algorithm]
             :end-before: [END configure_sweep_job_random_sampling_algorithm]
@@ -97,7 +96,6 @@ class GridSamplingAlgorithm(SamplingAlgorithm):
 
     .. admonition:: Example:
 
-
         .. literalinclude:: ../../../../../../samples/ml_samples_sweep_configurations.py
             :start-after: [START configure_sweep_job_grid_sampling_algorithm]
             :end-before: [END configure_sweep_job_grid_sampling_algorithm]
@@ -123,7 +121,6 @@ class BayesianSamplingAlgorithm(SamplingAlgorithm):
     """Bayesian Sampling Algorithm.
 
     .. admonition:: Example:
-
 
         .. literalinclude:: ../../../../../../samples/ml_samples_sweep_configurations.py
             :start-after: [START configure_sweep_job_bayesian_sampling_algorithm]