+++ sdk/monitor/azure-monitor-opentelemetry/tests/autoinstrumentation/test_configurator.py
@@ -3,15 +3,15 @@ from unittest import TestCase
 from unittest.mock import patch
 
 from azure.core.tracing.ext.opentelemetry_span import OpenTelemetrySpan
-from azure.monitor.opentelemetry.autoinstrumentation._configurator import (
+from azure.monitor.opentelemetry._autoinstrumentation.configurator import (
     AzureMonitorConfigurator,
 )
 
 
 class TestConfigurator(TestCase):
-    @patch("azure.monitor.opentelemetry.autoinstrumentation._configurator._is_attach_enabled", return_value=True)
+    @patch("azure.monitor.opentelemetry._autoinstrumentation.configurator._is_attach_enabled", return_value=True)
     @patch(
-        "azure.monitor.opentelemetry.autoinstrumentation._configurator.AzureDiagnosticLogging.enable"
+        "azure.monitor.opentelemetry._autoinstrumentation.configurator.AzureDiagnosticLogging.enable"
     )
     def test_configure(self, mock_diagnostics, attach_mock):
         configurator = AzureMonitorConfigurator()
@@ -20,9 +20,9 @@ class TestConfigurator(TestCase):
             configurator._configure()
         mock_diagnostics.assert_called_once()
 
-    @patch("azure.monitor.opentelemetry.autoinstrumentation._configurator._is_attach_enabled", return_value=False)
+    @patch("azure.monitor.opentelemetry._autoinstrumentation.configurator._is_attach_enabled", return_value=False)
     @patch(
-        "azure.monitor.opentelemetry.autoinstrumentation._configurator.AzureDiagnosticLogging.enable"
+        "azure.monitor.opentelemetry._autoinstrumentation.configurator.AzureDiagnosticLogging.enable"
     )
     def test_configure_preview(self, mock_diagnostics, attach_mock):
         configurator = AzureMonitorConfigurator()