+++ sdk/search/azure-search-documents/azure/search/documents/indexes/_search_index_client.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-from typing import Any, Dict, List, Union, Optional
+from typing import Any, Dict, List, Union, Optional, MutableMapping
 
 from azure.core import MatchConditions
 from azure.core.credentials import AzureKeyCredential, TokenCredential
@@ -137,7 +137,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         return SearchIndex._from_generated(result)  # pylint:disable=protected-access
 
     @distributed_trace
-    def get_index_statistics(self, index_name: str, **kwargs: Any) -> Dict:
+    def get_index_statistics(self, index_name: str, **kwargs: Any) -> MutableMapping[str, Any]:
         """Returns statistics for the given index, including a document count
         and storage usage.
 
@@ -182,7 +182,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         error_map, access_condition = get_access_conditions(index, match_condition)
         kwargs.update(access_condition)
         try:
-            index_name = index.name
+            index_name = index.name  # type: ignore
         except AttributeError:
             index_name = index
         self._client.indexes.delete(index_name=index_name, error_map=error_map, **kwargs)
@@ -391,7 +391,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         error_map, access_condition = get_access_conditions(synonym_map, match_condition)
         kwargs.update(access_condition)
         try:
-            name = synonym_map.name
+            name = synonym_map.name  # type: ignore
         except AttributeError:
             name = synonym_map
         self._client.synonym_maps.delete(synonym_map_name=name, error_map=error_map, **kwargs)
@@ -537,7 +537,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         error_map, access_condition = get_access_conditions(alias, match_condition)
         kwargs.update(access_condition)
         try:
-            alias_name = alias.name
+            alias_name = alias.name  # type: ignore
         except AttributeError:
             alias_name = alias
         self._client.aliases.delete(alias_name=alias_name, error_map=error_map, **kwargs)